Added method for covering of a tile by convex polygons
This is the sort of thing you can always calculate when you need to, but the
algorithms are not dirt-simple, and in our case we can hand-code a covering
very easily.

Now, *why* would you want a covering of a tile by convex polygons, you might
ask? Because it makes determining whether two tiles intersect so much easier,
that's why.